Add feature guards to grid daemon
The grid daemon's feature set is now guarded by the following features:

    - pike
    - product
    - schema
    - location
    - track-and-trace

When these features are enabled the corresponding rest api routes will
be active, the app auth handler will upload corresponding smart
contracts to created circuits and the database handler will listen for
corresponding events. Pike, product, schema, and location are considered
stable features and will be included by default or if the daemon is
compiled with the stabled flag. Track and trace is considered
experimental and will only be included if the daemon is compiled with
the experimental flag.

When only location and product are enabled, pike and schema will be
included as well because those features are  required to properly use
product and location. Similarly, if schema is enabled pike will also be
included for the same reason.

Signed-off-by: Ryan Banks <rbanks@bitwise.io>