Add feature guards to CLI
The grid cli feature set is now guarded by the following features:

- pike
- product
- schema
- location

When these features are enabled the corresponding set of cli commands
will be enabled. Pike, product, schema, and location are considered
stable features and will be included if the cli is compiled with the
stable flag.

When only location and product are enabled, pike and schema will be
included as well because those features are required to properly use
product and location. Similarly, if schema is enabled pike will also be
included for the same reason.

Signed-off-by: Ryan Banks <rbanks@bitwise.io>