Add feature guards to sdk
Adds the following features to the grid sdk

    - pike
    - product
    - schema
    - location
    - track-and-trace

Pike, product, schema, and location are considered
stable features and will be included if the sdk is compiled with the stabled
flag. Track and trace is considered experimental and will only be included if
the sdk is compiled with
the experimental flag.

When only location and product are enabled, pike and schema will be
included as well because those features are  required to properly use
product and location. Similarly, if schema is enabled pike will also be
included for the same reason.

Signed-off-by: Ryan Banks <rbanks@bitwise.io>