Modify displayName in oauth-login sapling
url decode 'displayName' in the oauth-login sapling. This change
is necessary for the situation where the account used for
logging in to splinter admin ui exists but does not have an
associated email address. In this case the value of displayName
will be the user's 'name' which could contain spaces.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>