Add `display_name` throughout Circuits Sapling
Updates the Circuits Sapling's circuit definition to include the
`display_name`, including the protobuf message.

This also includes updates to the table of circuits to include a `Display name`
column, which allows sorting by the `display_name`.

Also adds an option to set a circuit's `display_name` when proposing a
circuit in the UI.

Signed-off-by: Shannyn Telander <telander@bitwise.io>