Update error handling in Circuits sapling
Updates how errors are initially handled, to pull out the error message,
rather than the entire error object, which did not render in the UI as a
human-readable error message.

Also updates the `CircuitDetail` page to display an error message if
encountered when attempting to fetch the specific circuit or proposal
from the Splinter daemon.

Signed-off-by: Shannyn Telander <telander@bitwise.io>