Consolidate addressing in sdk
This consolidates the addressing functionality for various Grid smart
contracts into corresponding addressing functions in the SDK.
Previously, this functionality was duplicated for each smart contract
resulting in a lot of extra code. Consolidating the addressing in the
SDK will also make any future updates to the addressing for a resource
much easier.

Signed-off-by: Davey Newhall <newhall@bitwise.io>