Include record_owner in has_permission check
This change includes the record owner in the API for has_permission.
It now includes the check to ensure that the agent exists and has the
same org as the owner of the record that the has_permission check
concerns. Previously this had to be done in the smart contract handler
logic.

Signed-off-by: Darian Plumb <dplumb@bitwise.io>