Update Schema smart contract to use new has_permission check
This commit updates the Schema smart contract's calls to has_permission
based on the new API. The call now includes the record owner, and
checks whether the agent exists. Several tests were removed that were
redundant with the unit tests already in the permissions module. The
error text returned with a false permission check was updated to be
more specific.

Signed-off-by: Darian Plumb <dplumb@bitwise.io>