Update add_role DB operation
This updates the add_role database operation to account for the addition
of the state_address association table. This operation now adds an entry
to that table that maps the role name and organization id to the role's
address in state. The update role operation does not need a similar
change at this time because role names and org IDs can't change.

Signed-off-by: Davey Newhall <newhall@bitwise.io>