Add sqlite tests to integration test and add chain commands
Adds sqlite tests to integration docker compose and chains together
commands using "&". Chaining commands together is necessary because
without chaining, the Jenkins build will pass as long as the last set of
tests pass, regardless as to whether or not the other test suits
failed leading to Jenkins falsely reporting that all the tests passeed.

Signed-off-by: Ryan Banks <rbanks@bitwise.io>