Allow redundance clones in store_state
This linting error occurs when the sdk is built as part of griddle,
which does not specify the various features within.

Instead of fixing this, simply marking it as allow, because store state
needs to be rewritten to use the store factory instead of pg/sqlite, and
this lint-allow can be addressed at the same time.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>