Add custom icon support to sidebar
This adds support for custom icons for a sapling in the sidebar. These
custom icons are specified by setting the image field in a sapling's
entry in the sapling manifest with a URL to the image. The custom image
takes precedence over an icon set with the icon field but at least one
of these fields is required.

Signed-off-by: Davey Newhall <newhall@bitwise.io>