Implement multiple file handler for product
Updates the product create CLI command to take any number of occurances
of --file. The resulting payloads will have all properties specified
across the files for each product ID.

For instance, you can submit an XML file with a GDSN trade item
definition in one file, and then submit a yaml file with the same GTIN
as the product ID to supply additional attributes. This would require a
custom schema to do.

This commit also removes the conflict between the "file" and "owner"
arguments. This is because GDSN inputs currently require the owner to be
supplied. A future commit should implement an alternate ID lookup on the
GS1 company prefix based on the GTIN of the product and reinstate this
conflict.

Signed-off-by: Darian Plumb <dplumb@bitwise.io>