Add last_updated field to gridd endpoints
Adds last_updaed field to product, location, agent, organization, role,
and schema endpoints. The last_updated timestamp is generated by the
database when the element is inserted into the database. The database
stores the last_updated as a TIMESTAMP object, but the rest api will
return date created as seconds since January 1, 1970.

Signed-off-by: Ryan Banks <rbanks@bitwise.io>