Switch libxml for libc validation implementation
The rust-libxml bindings library did not handle validation errors
properly. Instead of reporting all of the validation errors for a file,
the library would return the correct number of errors but all of the
error messages would be the same. This issue seems to be due to the
implementation of the StructuredError handler in the rust-libxml
bindings library.

This commit removes the bindings library and just implements bindings
for the specific libxml2 functions needed for validation.

Signed-off-by: Darian Plumb <dplumb@bitwise.io>