Update Commit store errors
This updates the commit store error module. This removes the
CommitEventError error type as it was a duplicate of CommitStoreError.
This also moves the From implementations of CommitStoreError to the
error.rs file to be in line with the best practices. Lastly, this
updates the CommitStore implementations to use the From implementations.

Signed-off-by: Davey Newhall <newhall@bitwise.io>