Remove "pub use" statements for stores
Previously, the modules containing the stores did "pub use" on the
stores. For example, you could do "use product::ProductStore". This
change removes this such tat you must "use
product::store::ProductStore".

The primary motivation for this change is to ensure there is only
a single correct way to import the stores. Since the store modules are
pub, it makes sense to remove "pub use". A secondary motivation is
simply to keep the compartmentalization clear.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>