Make all diesel modules pub(crate)
Ideally, we would make all the diesel modules private; however, the
migrations module breaks compartmentalization by reaching inside the
schema modules.

This uses "pub use" for diesel store implementations. This makes sense
primarily because it is the only public thing exported (and nothing else
should be), and we are already scoped to the associated store module.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>