Update CLI PO create to include not-draft option
The Man page specifies that there would be both a draft - which is by
default, and a not-draft flag. Updated this implementation to match the
man-page, because draft is a flag and is set to true by default, there
would be no way of setting it to false. Also draft and not-draft is
the convention used for other components. This method is not yet
implemented so it should have no impact on the application.

Signed-off-by: Kevin Johnson <kevin_johnson@cargill.com>