Update rest api tests to use :memory: connection
This updates the rest api tests to use an in-memory sqlite db. This
offers a couple benefits. One, this "resets" the database between tests
so that the db does not need to be cleared manually. Two, this does not
create any database files that would need to be manually cleaned up
after the tests.

Signed-off-by: Davey Newhall <newhall@bitwise.io>