Fix rest API URL handling in `run` methods
This change creates an `Endpoint` object from the `GridConfig`'s
`endpoint` immediately as it is passed into the daemon's `run` methods
for Splinter and Sawtooth backends. This allows for the backend rest
API URL to be correctly formatted as it is returned by the
config object with a `splinter:` or `sawtooth:` prefix to allow the
daemon to start correctly. This fixes the issue where the `splinter:`
and `sawtooth:` prefixes are being passed in with the URL, returning
an error as this creates an invalid URL scheme.

Signed-off-by: Shannyn Telander <telander@bitwise.io>