Add get_current_service_commits
This change adds a method to the CommitStore to provide the current
commits for all services.

It also adds the implementation for the DieselCommitStore, supplying the
query via an operation trait.  The operation implementation itself
requires using a raw SQL query, as the query cannot be generated using
the diesel DSL (as of diesel 1.4.x releases).

It is guarded behind the experimental feature
"commit-store-service-commits".

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>