Remove public fields from Pike Organization
This change removes the public fields from the PikeStore's Organization
struct. Public accessor methods are implemented to return references to
the Organization fields. This also requires changing the `From`
implementation for `OrganizationMetadata` and `AlternateId`, as these
objects are now returned as references from the Organization. This
change also includes updating how the Organization is built within the
db_handler in the Grid daemon, using the OrganizationBuilder in place of
directly assigning public fields.

Signed-off-by: Shannyn Telander <telander@bitwise.io>