Remove public fields from Pike Alternate ID
This change removes the public fields from the PikeStore's Alternate ID
struct. These public fields are replaced by public accessor methods.

This also updates the construction of an Alternate ID with the builder
implemented for this struct. This is necessary as the fields are no
longer public, and may not be assigned directly, requiring the builder.

Signed-off-by: Shannyn Telander <telander@bitwise.io>