Add experimental support for Cylinder JWTs
This adds support for using Cylinder JWTs as an Auth header in requests
to submit batches. This new functionality is guarded behind the
experimental `cylinder-jwt-support` feature. This creates the Signer in
the the `run_splinter` function which is then passed through to the
REST API when that is set up.

Signed-off-by: Davey Newhall <newhall@bitwise.io>