Remove admin-keygen from CLI
This change removes the `admin` module from the CLI actions and deletes
the corresponding admin-keygen feature. This feature was not used by the
CLI and the `admin-keygen` command has been replaced by the `keygen`
command and its related options.

This change also removes a call to the removed command from the Splinter
example's docker compose file, as this command is already replaced by
the `keygen --system` command.

Signed-off-by: Shannyn Telander <telander@bitwise.io>