Add Cylinder JWT support to app auth handler
This change adds authorization, in the form of a Cylinder JWT, to the
app auth handler's request sent to Splinter. This is required by
Splinter 0.5. This allows the Grid daemon to start with the Splinter 0.5
backend.

This change also updates the signer used in the app auth handler from a
Sawtooth and Transact Signer to a Cylinder Signer. This supports creating
a JWT token from this signer as required for authorization.

Signed-off-by: Shannyn Telander <telander@bitwise.io>