Support Splinter 0.5 in docker compose
This change updates the Splinter dep in the docker-compose file for the
Splinter example from 0.4 to 0.5. This also updates the arguments in the
Splinter-specific containers to match the updated arguments for 0.5.

This change also adds volumes to hold the Grid daemon key, generated in
the Grid container and copied over to the Splinter container. This
allows the key to be put into the Splinter `allow_keys` file, which
enables this key to make authorized requests to Splinter to allow the
Grid daemon to successfully start using Splinter v0.5.

This change also updates the Scabbard-CLI package used in the
docker-compose file from v0.4 to v0.5.

Signed-off-by: Shannyn Telander <telander@bitwise.io>