Serialize splinter commit events
This change serializes the handling of commit events into the database.
It ensures that all commit events are handled in the order that the are
initially received, as well as no overlapping commit events are handled
simultaneously.

It fixes an issue where duplicate commit events may be received at
roughly the same time, and both will commit, as there is a flaw in the
way a transaction is generated.

This commit may be rolled back if the transaction issue is resolved.
However, that problem requires considerably more effort to correct.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>