Add and implement Transactional/InContext store factories
This adds the `TransactionalStoreFactory` and `InContextStoreFactory`
traits. The `TransactionalStoreFactory` trait contains functionality to
begin a DB transaction on a connection while the `InContextStoreFactory`
trait contains functionality to commit or rollback that transaction.
This allows for Grid to process DB events correctly for state delta
export. This commit also contains implementations of these traits for
both Postgres and Sqlite store factories.

Signed-off-by: Davey Newhall <newhall@bitwise.io>