Add action stubs for Purchase Order smart contract
This change adds unimplemented functions for the Purchase Order
smart contract's transaction handler actions. Also, this adds the
handling in the `apply` function to determine a payload's action.

This change also adds a dependency on the `pike` feature to the
`purchase-order` feature in the Grid SDK. This allows the Purchase Order
smart contract to use the Pike-related functionality from the SDK.

Signed-off-by: Shannyn Telander <telander@bitwise.io>