Rename attribute to match original source
This change renames the `accepted_version_number` within the Purchase
Order state struct and the payload to update a Purchase Order to
`accepted_version_id`. This name better matches the value that is
assigned to this attribute, as it is the ID of the purchase order
version that has been accepted.

Signed-off-by: Shannyn Telander <telander@bitwise.io>