Update has permission check
This updates the has_permission check to properly check permissions for
inherited roles. Previously, the record owner had to be the same as the
signing agent's organization even if the agent had properly inherited
permissions from the record owner. This updates it so it checks allowed
organizations for the applicable role.

Signed-off-by: Davey Newhall <newhall@bitwise.io>