Add workflow-aware has_permission check
This adds a new version of the has_permission check that is used in
other smart contracts. This check is workflow-aware by checking against
workflow permissions and transition permissions. This is used in the
same way as the original has_permission check. Agents will need to have
Pike roles with permissions that map to the names of Workflow permission
aliases for the given workflow.

Signed-off-by: Davey Newhall <newhall@bitwise.io>