Update PO workflow
This updates the PO workflow and subworkflows to match the updated
version in the RFC. This also updates the names of the workflow
permission aliases because they will map to pike permissions and using
a "." separated format was confusing because pike roles also use that
syntax. This also removes several permission aliases in the subworkflows
as they are no longer relevant with the change to GS1 XML PO data.

Signed-off-by: Davey Newhall <newhall@bitwise.io>