Add workflow aware permission check method
This change adds a method to the `PermissionChecker` struct to be used
to validate an agent's permissions, specifically regarding a workflow.

This change updates the previously defined
`check_permission_with_workflow` to work within the `PermissionChecker`
struct, requiring removing the `perm_checker` argument and using the
`PermissionCheckerError`.

This also removes the `check_permission_with_workflow` from the Purchase
Order smart contract's handler module, including tests which have been
updated and moved to the correct module.

Signed-off-by: Shannyn Telander <telander@bitwise.io>