Update PO `uuid` to `uid`
This change updates the Purchase Order's `uuid` field to `uid`. The ID
is not validated to be a traditional UUID, so `uid` is a better term for
this field.

This change also updates the corresponding protocol code and
the corresponding Purchase Order state code to refer to the `uuid` as
`uid`.
Signed-off-by: Shannyn Telander <telander@bitwise.io>