Make `accepted_version_id` optional in SDK
This change makes the accepted version number within a purchase order
optional throughout the SDK. As this value will not be defined until a
user submits an update to a purchase order to move the specific version
to "accepted", this value should have the ability to be `None` rather
than an empty string.

This change updates the `accepted_version_id` field in the Purchase
Order store struct to be an optional value. This also updates the
corresponding `accepted_version_number` in the purchase order protobuf
message and protocol module to be optional.

Signed-off-by: Shannyn Telander <telander@bitwise.io>