Add `workflow_type` to po smart contract
This change adds consideration for the `workflow_type` to the purchase
order smart contract. This adds an implementation of the `Display` trait
to the `POWorkflow` struct to allow the po smart contract to extract the
`workflow_type` that is now added to the purchase order state object.
This change also updates the associated tests.

Signed-off-by: Shannyn Telander <telander@bitwise.io>