Add permissions for the `po::draft` alias
This change updates the `create` workflow state to add permissions for
the `po::draft` PermissionAlias within the `SystemOfRecord` purchase
order workflow. This allows an agent with the `po::draft` workflow
permission to create a purchase order version as a draft and transition
it to the `editable` workflow.

This also adds transitions within the `editable` workflow state to allow
an agent to transition a purchase order version from the `editable`
state to one of the `editable`, `cancelled`, `declined`, or `review`
states.

Signed-off-by: Shannyn Telander <telander@bitwise.io>