Fix handling of protobuf optional field
This change fixes how the `CreateVersionPayload` is being converted by
the purchase order protocol code. Previously, this value was being
pulled from the protobuf, but as it is optional, the value pulled from
the protobuf must first be validated as non-empty. If it is empty, this
field is set to `None`. If not, the `CreateVersionPayload` is converted
into a native struct using it's `from_proto` method.

This fixes an issue where the payload may be set with empty values from
the protobuf.

Signed-off-by: Shannyn Telander <telander@bitwise.io>