Create JWT at Splinter run-time
This change updates the `run_splinter` method to generate a Cylinder JWT
at run-time.

This also updates the Splinter backend client to hold the authorization
as a field, added in the `new` method. This also updates the backend
client's `submit_batches` to use the `authorization` field of the
client, rather than generating a JWT within the method.

Signed-off-by: Shannyn Telander <telander@bitwise.io>