Update `id` to be `BIGSERIAL` in pg migrations
This change updates the `id` field to `BIGSERIAL`. This field is used as
an incrementing id when inserting new records into a PostGres database.

Previously, with the type `INTEGER`, this field would not be
automatically set when being inserted into a PostGres database and would
error because of this null value.

Signed-off-by: Shannyn Telander <telander@bitwise.io>