Update schema loading for Order.xsd
This change updates the loading of the schema to be done in the
directory where schemas are specified. This directory is able to be set
by passing in the directory that contains the Product or Purchase Order
schemas. This argument is provided to the `load_schema` and
`validate_xml` functions.

This adds environment variables to the cli module, to allow for
defining the directory passed to the `validate_xml` function.
This change adds a `GRID_ORDER_SCHEMA_DIR` and a
`GRID_PRODUCT_SCHEMA_DIR` environment variables to allow for users to
define each of these directory paths, which will allow for the use of
validation files outside of the SDK.

Signed-off-by: Shannyn Telander <telander@bitwise.io>