Update purchase order test module
This change makes a few minor updates to support the tests to be written
for the `update version` action. Mainly, this adds a public key for an
agent with the `po::draft` permission. This allows for imitating
multiple agents with separate and individual permissions, rather than one
public key overloaded with multiple permissions.

This change also updates the function used to create a purchase order,
as currently this function creates a purchase order with a draft
version. This change updates this function to create a purchase order with
a version in the `proposed` state and updates previous uses of this
function to use a combination of functions to create a purchase order
with a draft version.

Signed-off-by: Shannyn Telander <telander@bitwise.io>