Update set version in state to consider updates
This change updates the `set_purchase_order_version` method in the
purchase order state module to consider if a purchase order version is
being updated. This means, when the po is fetched from state, the method
will now check to see if a version with the same ID exists in the list
and then it is removed and re-inserted with the updated version.
Otherwise, a new version is simply added to the end of the purchase
order's `versions` list.

Signed-off-by: Shannyn Telander <telander@bitwise.io>