Validate self workflow state in `can_transition`
This change updates the workflow state module's `can_transition` method,
which validates whether an agent has the correct permissions to
transition to a new workflow state, to check the workflow state's name.
This covers a case in which an agent is not attempting to transition a
workflow, but rather update an object.

Previously, the `can_transition` method would return false if it did not
have it's own name within it's transitions. This will now check if we are
staying in the same workflow state and return `true`, as we are not
transitioning.

This also required updating the `test_workflow_state` test to rename the
workflow state used to have a different name from the transitions. The
test validates that a permission alias is unable to transition if the
desired state is not within the alias' available `transitions`.

Signed-off-by: Shannyn Telander <telander@bitwise.io>