Add `alternate_id` lookup to PO store operations
This updates the `purchase_order_store` db operations to support UID
lookup from PO alternate IDs. This change makes it so any ID in the
alternate ID format (`<id_type>:<id>`) that is passed to an operation is
first looked up in the alternate ID table and converted to its UID
counterpart.

Signed-off-by: Davey Newhall <newhall@bitwise.io>