Reorganize client module
This reorganizes the client module. Previously, the module placed the
client and reqwest client files next to each other separated by feature.
Now, the module is split to have the client files together and the
reqwest specific files grouped to simplify the config. This also splits
out and renames the "DTO" objects into separate `data::*` modules.

Signed-off-by: Davey Newhall <newhall@bitwise.io>