Get global arg values at lowest subcommand
Implements a fix regarding values of global options being
unavailable to subcommands.

While clap propagates global argument values down to the
subcommand level, they are not available in intermediate levels.
This is in contrast to how we must define them in the app.

Previously, the CLI tried to access global arg values at the level
that they were defined. This fix gets the values at the lowest
subcommand level.

Signed-off-by: Chris Eckhardt <eckhardt@bitwise.io>