Fix po update cli to use po ids from po
Previously the po update command did not work, as it used an
unimplemented REST API client method to fetch alternate ids. The client
now correctly parses and returns the alternate ids from the server when
the po is fetched, so this change updates the po update command to use
that instead.

Signed-off-by: Lee Bradley <bradley@bitwise.io>