Update "create po" action to handle constraints
This change updates the "create po" action in the purchase order smart
contract to concern itself with workflow constraints. This adds further
validation when a user attempts to create a purchase order.

This change also adds a test to validate a purchase order may not be
created if the desired workflow state has a constraint it does not
fulfill.

Signed-off-by: Shannyn Telander <telander@bitwise.io>