Update mid-txn errors to be InvalidTransaction
This change updates multiple occurrences of `InternalError` being
returned to the user from the purchase order smart contract handler.
These errors are specific to the transaction being invalid, and therefore
should result in the `InvalidTransaction` error rather than `IternalError`.
`InternalError` refers to errors that are unexpected and system-level;
the errors updated in this commit pertain to the validity of the
transaction.

Signed-off-by: Shannyn Telander <telander@bitwise.io>