Add route and client handler for `get_latest_revision_id`
This adds the functionality corresponding to the
`get_latest_revision_id` DB operation to the REST API and PO client.
This consists of adding a route,
`/purchase_order/{id}/version/{version_id}/revision/latest`, to the
REST API and adding functionality to the client to handle fetching from
that endpoint.

Signed-off-by: Davey Newhall <newhall@bitwise.io>