Add "streaming" functionality to list operations
This adds functionality to the purchase order client to "stream" results
back. This is done by creating an iterator that holds results in a cache
and uses the paging data from the result to add new data if there are
additional pages. This also required significant updates to how the cli
handled printing tabulated results.

Signed-off-by: Davey Newhall <newhall@bitwise.io>