Fix revision count in `grid po list`
Adds filter to select only revisions of the same purchase order
when listing purchase orders. Previously, if versions of different
purchase orders had the same version ID, the revision count would
include all revisions of versions with that name.

Note that revision count is not currently displayed in `grid po
list`. This change is for consistency.

Signed-off-by: Chris Eckhardt <eckhardt@bitwise.io>