Update permission check for workflow entry
This change updates the `check_permission_with_workflow` method by
renaming it to `check_permission_within_workflow` as this method has the
ability to validate a user's permission to alter an object within the
workflow. With the addition of the `StartWorkflowState`, the entry
point of a subworkflow, the permission checker has the ability to
validate a user's permission to create an object within a workflow.
This change adds a `check_permission_to_enter_workflow` method to the
`PermissionChecker` to validate a user's permissions within a
`StartWorkflowState`.

Signed-off-by: Shannyn Telander <telander@bitwise.io>