Add `StartWorkflowState` to `SubWorkflow`
This change updates the `SubWorkflow` struct to hold a single
`start_state`, which is represented by the `StartWorkflowState` struct.
This field is now used rather than `starting_states`, which previously
held the names of workflow states an object may begin a subworkflow at.
Now, this field more specifically defines the user permissions and valid
transitions for an object before it has been created or entered the
workflow.

Signed-off-by: Shannyn Telander <telander@bitwise.io>