Update po smart contract to use start state
This change updates the built-in purchase order workflows to use a
`start_state` rather than directly defining a "create" state. The start
state defines the permissions for entering a workflow and are not valid
`WorkflowState`s, therefore the "create" state is now the `start_state`
of the SubWorkflow.

This also updates the smart contract to use the correct permission
checker method when creating an object. The `PermissionChecker`'s
`check_permission_to_enter_workflow` method is now used as this
specifically inspects a `StartWorkflowState` to determine if the object
may be created within the workflow.

This change also updates an affected test, to remove any direct
references to a "create" workflow state when testing the ability to
create a purchase order.

Signed-off-by: Shannyn Telander <telander@bitwise.io>