Update "create version" action in smart contract
This updates the `create_version` function in the purchase order smart
contract that handles the creation of a purchase order version. This
change updates the workflow state used to validate a user's permission
to create an object and enter it into a workflow and the general layout
of the function to match the layout of the other actions. This also
updates affected tests.

Signed-off-by: Shannyn Telander <telander@bitwise.io>