Remove actix-web default features
This change removes the default features from the actix-web dependency
which removes the transient dependency on brotli for compression (as well as gzip).

This fixes a compilation issue on linux systems where there is an impedence
mismatch between actix-http 2.x (pulled in by actix-web-3) and awc 0.2.x
 (pulled in by the splinter/events feature).

As compression is not used in any of the clients, it is acceptible to remove
these transient dependencies.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>