Check if version exists in create version command
This change updates how the revision ID is fetched for a version being
created. As this version is being created, the revision ID should be '1'
rather than the most recent revision ID of the version. The most recent
revision ID should not exist, as a version being created should not
already exist.

Signed-off-by: Shannyn Telander <telander@bitwise.io>