Remove `hex-literal` library
Remove the hex-literal library in favor of runtime hex interpretation,
which is available as a standard Rust feature. This is to prevent
the buildup of superfluous libraries and resulting technical debt.

This change also removes the abstraction of the Hash generic from the
xsd download utility, just using a &str instead.

Signed-off-by: Lee Bradley <bradley@bitwise.io>