Cache Grid XML in Github Actions
Adds a GHA workflow that runs each day at 10 UTC to create an dependency cache
for Grid builds. This is necessary because the upstream provider has a
regular maintenance window that makes these dependencies unavailable for an
uncertain amount of time which causes disruptions in development.

This Action also contains a test which attempts to restore and consume
dependencies from the cache. This tests that the cache is available and as
part of restoring, that the dependency hashes match what we expect.

Signed-off-by: Ryan Beck-Buysse <rbuysse@bitwise.io>