Add `proxy_get` endpoint to Griddle
This change adds the `proxy_get` endpoint to the Griddle REST API. This
also adds the arguments required to produce a `ProxyClient` to use for
the proxy endpoint.

This change also removes starting the `BatchProcessor` from the
griddle docker-compose file. The `BatchProcessor` is a previously
implemented component that will be replaced.

Signed-off-by: Shannyn Telander <telander@bitwise.io>