Remove `[macro_use]` from serde_json import
This change removes the `#[macro_use]` decorator from the serde_json
import in sdk/lib.rs and directly import the `json` macro in the module
it is used. This is required because the `proxy-client` feature that
also uses `serde_json` does not use macros from this crate, which
produced a warning when compiling this feature.

Signed-off-by: Shannyn Telander <telander@bitwise.io>