Update batch tracking migrations
This updates the database schema for the batch tracking tables by
adding batch/transaction ID columns. These columns are intended to be
the primary keys for a table and consist of a hash of the batch/txn
header and the service ID. Columns for the service ID and batch/txn
headers are also included for deserialization sake. This is a necessary
change because diesel does not fully support composite keys.
Specifically, it does not implement the `BelongsToDsl` trait for
composite keys and they do not intend to implement this functionality,
as stated in https://github.com/diesel-rs/diesel/issues/1413.

Signed-off-by: Davey Newhall <newhall@bitwise.io>