Undo batch_tracking migrations key changes
Previously, the migrations for the batch tracking tables were updated
to use a surrogate key composed of the id and service id due to Diesel's
incomplete support for composite foreign keys. This change switches the
primary key back to a composite primary key as well as the related
foreign keys and operations that need to filter based on those foreign
keys will use raw sql via the `sql_query` function.

Signed-off-by: Davey Newhall <newhall@bitwise.io>