Add intitial v2 module for transaction payloads
This change adds an initial module for the V2 transaction payloads
used to deserialize HTTP requests into specific contract actions.

This also adds the `TransactionPayload` trait in the `payloads` module.
The contract actions represented in this module will implement the
trait, allowing for easier conversion from JSON to signed Batch.

The v2 module is behind an experimental
`rest-api-resources-batch-tracking` feature.

Signed-off-by: Shannyn Telander <telander@bitwise.io>